requirements.txt
newspaper3k==0.2.8
trafilatura==1.8.1
playwright==1.46.0
beautifulsoup4==4.12.3
...
lxml
...
langchain==0.1.20
langchain-community==0.0.29
sentence-transformers==2.7.0
chromadb==0.4.24
groq==0.8.0
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.2